django_project/requirements.txt
CacheControl==0.12.10
cachy==0.3.0
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
debugpy==1.2.1
distlib==0.3.3
docopt==0.6.2
filelock==3.3.2
future==0.18.2
html5lib==1.1
jedi==0.15.2
jeepney==0.7.1
keyring==21.8.0
lockfile==0.12.2
msgpack==1.0.2
packaging==20.9
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
Pillow==10.0.0
pipreqs-amasad==0.4.10
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==0.13.1
poetry==1.1.11
poetry-core==1.0.7
ptyprocess==0.7.0
pycparser==2.21
pyflakes==2.1.1
PyJWT==2.8.0
pylev==1.4.0
pyparsing==3.0.5
python-dotenv==1.0.0
python-jsonrpc-server==0.3.2
python-language-server==0.31.10
reportlab==4.0.4
requests-toolbelt==0.9.1
rope==0.18.0
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
toml==0.10.2
tomlkit==0.7.2
ujson==1.35
virtualenv==20.10.0
webencodings==0.5.1
yapf==0.31.0
yarg==0.1.9

